endpoint added for advance and advance logic implamented

Related work items: #122, #133

+++ Core/HrApp.Application/CQRS/Advance/Commands/DeleteAdvanceCommand.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace HrApp.Application.CQRS.Advance.Commands
 {
-    public class DeleteAdvanceCommand : IRequest<ServiceResponse<int>>
+    public class DeleteAdvanceCommand : IRequest<ServiceResponse<decimal>>
     {
         public int Id { get; set; }
